Consolidate programService index file
